Refactor hash fixtures
