+++ p2panda-rs/src/test_utils/fixtures/new_fixtures/mod.rs
@@ -1,5 +1,7 @@
 mod entry_fixtures;
+mod hash_fixtures;
 mod operation_fixtures;
 
 pub use entry_fixtures::*;
+pub use hash_fixtures::*;
 pub use operation_fixtures::*;
